优化删除远端服务器上的好友

+++ src/cn/ucai/chatuidemo/activity/MainActivity.java
@@ -582,30 +582,25 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 		@Override
 		public void onContactDeleted(final List<String> usernameList) {
 			// 被删除
-			Map<String, User> localUsers = ((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList();
-
 			String currentUserName = SuperWeChatApplication.getInstance().getUserName();
-			List<String> toDelUserName = new ArrayList<String>();
-			for (String username : usernameList) {
-				localUsers.remove(username);
-				toDelUserName.add(username);
-				userDao.deleteContact(username);
-				inviteMessgeDao.deleteMessage(username);
-			}
-			for (final String name : toDelUserName) {
+			for (final String username : usernameList) {
 				final OkHttpUtils2<Result> utils = new OkHttpUtils2<Result>();
 				utils.setRequestUrl(I.REQUEST_DELETE_CONTACT)
 						.addParam(I.Contact.USER_NAME,currentUserName)
-						.addParam(I.Contact.CU_NAME,name)
+						.addParam(I.Contact.CU_NAME,username)
 						.targetClass(Result.class)
 						.execute(new OkHttpUtils2.OnCompleteListener<Result>() {
 							@Override
 							public void onSuccess(Result result) {
-								Map<String, UserAvatar> userMap = SuperWeChatApplication.getInstance().getUserMap();
-								List<UserAvatar> userList = SuperWeChatApplication.getInstance().getUserList();
-								UserAvatar u = userMap.get(name);
-								userList.remove(u);
-								userMap.remove(name);
+								if (result.isRetMsg()) {
+									((DemoHXSDKHelper) HXSDKHelper.getInstance()).getContactList().remove(username);
+									UserAvatar u = SuperWeChatApplication.getInstance().getUserMap().get(username);
+									SuperWeChatApplication.getInstance().getUserList().remove(u);
+									SuperWeChatApplication.getInstance().getUserMap().remove(username);
+									userDao.deleteContact(username);
+									inviteMessgeDao.deleteMessage(username);
+									sendStickyBroadcast(new Intent("update_contac_list"));
+								}
 							}
 
 							@Override
